+++ src/Unit Tests/ValidationTests.cs
@@ -1,17 +1,11 @@
-﻿using DigitalProduction.Maui.Converters;
-using System.Diagnostics;
+﻿using Data.Translation.Validation;
 
 namespace DigitalProduction.UnitTests;
-
-public class ConcreteEnumDescription : EnumToDescriptionConverter<TestingType>
-{
-	public ConcreteEnumDescription() { }
-}
 
 /// <summary>
 /// Test cases the the Mathmatics namespace.
 /// </summary>
-public class ConverterTests
+public class ValidationTests
 {
 	#region Members
 	#endregion
@@ -22,81 +16,36 @@ public class ConverterTests
 	/// Test to convert a relative path to an absolute path.
 	/// </summary>
 	[Fact]
-	public void EnumToDescriptionConverterTest()
+	public void IsNumericRuleTest()
 	{
-		string errorMessage									= "Enum description test failed.";
-		EnumToDescriptionConverter<TestingType> converter	= new();
+		string errorMessage	= "IsNumericRule test failed.";
 
-		string stringResult = (string)converter.Convert(TestingType.Type1, typeof(TestingType), null, System.Globalization.CultureInfo.CurrentCulture);
-		Assert.True(stringResult == "Type 1", errorMessage);
+		// Basic checks that non-numeric is false and numeric is true.
+		IsNumericRule rule	= new();
 
-		TestingType result = (TestingType)converter.ConvertBack("Type 1", typeof(TestingType), null, System.Globalization.CultureInfo.CurrentCulture);
-		Assert.True(result == TestingType.Type1, errorMessage);
-	}
+		Assert.False(rule.Check(null), errorMessage);
+		Assert.False(rule.Check(""), errorMessage);
+		Assert.False(rule.Check("A"), errorMessage);
 
-	/// <summary>
-	/// Test to convert a relative path to an absolute path.
-	/// </summary>
-	[Fact]
-	public void ConcreteEnumToDescriptionConverterTest()
-	{
-		string errorMessage					= "Enum description test failed.";
-		ConcreteEnumDescription converter	= new();
+		Assert.True(rule.Check("01"), errorMessage);
+		Assert.True(rule.Check("1.2"), errorMessage);
 
-		string stringResult = (string)converter.Convert(TestingType.Type1, typeof(TestingType), null, System.Globalization.CultureInfo.CurrentCulture);
-		Assert.True(stringResult == "Type 1", errorMessage);
-
-		TestingType result = (TestingType)converter.ConvertBack("Type 1", typeof(TestingType), null, System.Globalization.CultureInfo.CurrentCulture);
-		Assert.True(result == TestingType.Type1, errorMessage);
-	}
-
-		/// <summary>
-	/// Test to convert a relative path to an absolute path.
-	/// </summary>
-	[Fact]
-	public void AreDebuggingConverterTest()
-	{
-		string errorMessage				= "AreDebugging description test failed.";
-		AreDebuggingConverter converter	= new();
-
-		bool solution	= false;
-		string mode		= "Release mode.";
-		#if DEBUG
-			solution	= true;
-			mode		= "Debug mode.";
-			Debug.WriteLine("Debug mode.");
-		#else
-			Debug.WriteLine("Release mode.");
-		#endif
-
-		bool result = (bool)converter.Convert(TestingType.Type1, typeof(TestingType), null, System.Globalization.CultureInfo.CurrentCulture);
-
-		// Create a string array with the lines of text
-		string[] lines = {
-			mode,
-			"Result: "+result.ToString(),
-			"Expected result: "+solution,
-			"Passed: "+(result==solution).ToString(),
-			converter.AssemblyName,
-			"Attributes size: "+converter.DebuggableAttributes.Count,
-			"Just in time tracking enabled: "+converter.IsJustInTimeTrackingEnabled
-		};
-		WriteDebugInfo(lines);
-
-		Assert.True(result == solution, errorMessage);
-	}
-
-	private void WriteDebugInfo(string[] lines)
-	{
-		string filePath = Path.Combine(DigitalProduction.Reflection.Assembly.LibraryPath!, "DebugMessage.txt");
-
-		using (StreamWriter outputFile = new StreamWriter(filePath))
-		{
-			foreach (string line in lines)
-			{
-				outputFile.WriteLine(line);
-			}
-        }
+		// Check minimum and maximums.
+		rule.MinimumValue = 1.5;
+		rule.MaximumValue = 10;
+		Assert.False(rule.Check("0.2"), errorMessage);
+		Assert.True(rule.Check("1.5"), errorMessage);
+		Assert.True(rule.Check("6"), errorMessage);
+		Assert.True(rule.Check("10"), errorMessage);
+		Assert.False(rule.Check("12"), errorMessage);
+		
+		// Check decimal places.		
+		rule.MinimumDecimalPlaces = 1;
+		rule.MaximumDecimalPlaces = 2;
+		Assert.False(rule.Check("2"), errorMessage);
+		Assert.True(rule.Check("2.1"), errorMessage);
+		Assert.True(rule.Check("2.12"), errorMessage);
+		Assert.False(rule.Check("2.123"), errorMessage);
 	}
 
 	#endregion
